# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API for phpMyFAQ 4.1",
    "description": "phpMyFAQ includes a REST API and offers APIs for various services like fetching the phpMyFAQ version or doing a search against the phpMyFAQ installation.",
    "contact": {
      "name": "phpMyFAQ Team",
      "email": "support@phpmyfaq.de"
    },
    "license": {
      "name": "Mozilla Public Licence 2.0",
      "url": "https://www.mozilla.org/MPL/2.0/"
    },
    "version": "3.1"
  },
  "servers": [
    {
      "url": "https://localhost",
      "description": "Local dockerized server"
    }
  ],
  "paths": {
...
    "/api/v3.1/search": {
      "get": {
        "tags": ["Public Endpoints"],
        "operationId": "getSearch",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The search term",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the results for the query string search term.",
            "content": {
              "application/json": {
                "schema": {},
                "example": "\n        [\n            {\n                \"id\": \"1\",\n                \"lang\": \"en\",\n                \"category_id\": \"15\",\n                \"question\": \"Why are you using phpMyFAQ?\",\n                \"answer\": \"Because it is cool!\",\n                \"link\": \"https://www.example.org/index.php?action=faq&cat=15&id=1&artlang=en\"\n            }\n        ]"
              }
            }
          },
          "404": {
            "description": "If the search returns no results",
            "content": {
              "application/json": {
                "schema": {},
                "example": []
              }
            }
          }
        }
      }
    },
    "/api/v3.1/searches/popular": {
      "get": {
        "tags": ["Public Endpoints"],
        "operationId": "getPopularSearch",
        "responses": {
          "200": {
            "description": "Returns the popular search terms for the given language provided by \"Accept-Language\"",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "\n        [\n            {\n                \"id\": 3,\n                \"searchterm\": \"mac\",\n                \"number\": \"18\",\n                \"lang\": \"en\"\n            },\n            {\n                \"id\": 7,\n                \"searchterm\": \"test\",\n                \"number\": 9,\n                \"lang\": \"en\"\n            }\n        ]"
              }
            }
          },
          "404": {
            "description": "If the popular search returns no results.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": []
              }
            }
          }
        }
      }
    },
...
    "/api/v3.1/update": {
      "post": {
        "tags": ["Endpoints with Authentication"],
        "operationId": "triggerUpdate",
        "responses": {
          "200": {
            "description": "Returns the new and updated phpMyFAQ version number as string.",
            "headers": {
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "4.0.0"
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated and/or does not have sufficient permissions.",
            "headers": {
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}